feat: support to dump the ssrID
Implement SSR for Web #52
