+++ .changeset/dull-moons-scream.md
@@ -1,0 +1,5 @@
+---
+"@lynx-js/web-core-server": patch
+---
+
+feat: support to dump ssrID
