+++ .changeset/eighty-trees-appear.md
@@ -1,0 +1,5 @@
+---
+"@lynx-js/offscreen-document": minor
+---
+
+remove `dumpHTMLString`
