+++ packages/web-platform/offscreen-document/src/webworker/index.ts
@@ -1,11 +1,13 @@
 // Copyright 2023 The Lynx Authors. All rights reserved.
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
+export { OffscreenDocument, _onEvent } from './OffscreenDocument.js';
 export {
-  OffscreenDocument,
-  _onEvent,
-  dumpHTMLString,
-} from './OffscreenDocument.js';
+  _attributes,
+  innerHTML,
+  _children,
+  OffscreenElement,
+} from './OffscreenElement.js';
 export type * from './OffscreenEvent.js';
 export type * from './OffscreenElement.js';
 export type * from './OffscreenCSSStyleDeclaration.js';
